requirements.txt
astroid==2.15.6
autopep8==2.0.2
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
packaging==23.1
pipenv==2023.7.23
platformdirs==3.9.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pylint==2.17.5
python-dotenv==1.0.0
SQLAlchemy==2.0.19
tomlkit==0.11.8
typing_extensions==4.7.1
virtualenv==20.24.2
virtualenv-clone==0.5.7
Werkzeug==2.3.6
wrapt==1.15.0
